port more tests
