+++ packages/compat/package.json
@@ -78,7 +78,6 @@
     "@types/resolve": "^1.20.0",
     "@types/semver": "^7.3.6",
     "@types/strip-bom": "^4.0.1",
-    "broccoli-stew": "^3.0.0",
     "ember-cli-htmlbars-3": "npm:ember-cli-htmlbars@3",
     "ember-cli-htmlbars-inline-precompile": "^2.1.0",
     "ember-engines": "^0.8.19",
