+++ tests/scenarios/package.json
@@ -40,6 +40,7 @@
     "broccoli-funnel": "^3.0.5",
     "broccoli-merge-trees": "^3.0.2",
     "broccoli-persistent-filter": "^3.1.2",
+    "broccoli-stew": "^3.0.0",
     "ember-bootstrap": "^5.0.0",
     "ember-cli": "~3.28.0",
     "ember-cli-4.4": "npm:ember-cli@~4.4.0",
